requirements.txt
pandas
numpy
matplotlib
seaborn
azure-storage-blob